add clip

+++ model/SST.py
@@ -57,7 +57,7 @@ class SST(nn.Module):
 
         batchSize = input.size(0)
 
-        featureMap = self.backbone(input)                                            # (BatchSize, Channel, imgSize, imgSize)
+        featureMap = self.backbone(input).float()                                         # (BatchSize, Channel, imgSize, imgSize)
 
         if featureMap.size(1) != self.imageFeatureDim:
             featureMap = self.changeChannel(featureMap)                              # (BatchSize, imgFeatureDim, imgSize, imgSize)
